added filtering by status

+++ app/controllers/jobs_controller.rb
@@ -5,7 +5,7 @@ require 'tempfile'
 
 class JobsController < ApplicationController
   unloadable
-  before_filter :require_admin, :except => [:index, :show, :register, :filter, :export_to_csv, :zip_some, :zip_all, :zip_filtered, :view_table]
+  before_filter :require_admin, :except => [:index, :show, :register, :filter, :filter_by_status, :export_to_csv, :zip_some, :zip_all, :zip_filtered, :zip_filtered_single]
   
   helper :attachments
   include AttachmentsHelper
@@ -43,8 +43,8 @@ class JobsController < ApplicationController
     sort_init 'created_at', 'desc'
     sort_update 'last_name' => "#{Applicant.table_name}.last_name",
                 'id' => "#{JobApplication.table_name}.id",
-                'submission_status' => "#{JobApplication.table_name}.submission_status",
-                'acceptance_status' => "#{JobApplication.table_name}.acceptance_status",
+                'review_status' => "#{JobApplication.table_name}.review_status",
+                'offer_status' => "#{JobApplication.table_name}.offer_status",
                 'created_at' => "#{JobApplication.table_name}.created_at"
     #sort_update %w(id submission_status, acceptance_status, created_at)
     @job_applications = @job.job_applications.find(:all, :order => sort_clause)
@@ -322,8 +322,15 @@ class JobsController < ApplicationController
       
     end
     
-    send_file @zip_file_path, :type => 'application/zip', :disposition => 'attachment', :stream => false
-    File.delete(@zip_file_path)
+    begin
+      send_file @zip_file_path, :type => 'application/zip', :disposition => 'attachment', :stream => false
+      File.delete(@zip_file_path)
+    rescue  
+      if File.file?(@zip_file_path)
+        File.delete(@zip_file_path)
+      end  
+      puts "Error sending file"
+    end
     
   end
   
@@ -430,8 +437,15 @@ class JobsController < ApplicationController
       end
     end
     
-    send_file @zip_file_path, :type => 'application/zip', :disposition => 'attachment', :stream => false
-    File.delete(@zip_file_path)
+    begin
+      send_file @zip_file_path, :type => 'application/zip', :disposition => 'attachment', :stream => false
+      File.delete(@zip_file_path)
+    rescue  
+      if File.file?(@zip_file_path)
+        File.delete(@zip_file_path)
+      end  
+      puts "Error sending file"
+    end
   end  
   
   
@@ -549,28 +563,10 @@ class JobsController < ApplicationController
           job_app_ids.delete(app_id)
         end  
       end 
+      
       @job_applications = JobApplication.find(:all, :conditions => ["job_id = ? and id in (?)", params[:job_id], job_app_ids])
     end   
   end
-  
-  def view_table
-    @job = Job.find(params[:job_id])
-    unless User.current.admin? || @job.is_manager?
-      flash[:error] = "You are not authorized to view this section."
-  		redirect_to('/') and return
-  	end
-    @apptracker = @job.apptracker
-    @job_applications = @job.job_applications.find(:all)
-    @job_application_custom_fields = @job.all_job_app_custom_fields
-    @applicant_fields = Applicant.column_names - ["id", "created_at", "updated_at"]
-    @custom = []
-    unless @job_application_custom_fields.empty?
-  		@job_application_custom_fields.each do |custom_field|
-  		  @custom << custom_field.name
-  		end
-  	end
-  	@columns = @applicant_fields + @custom
-  end  
   
   def zip_filtered
     @job = Job.find(params[:job_id])
@@ -678,8 +674,150 @@ class JobsController < ApplicationController
 
     end
 
-    send_file @zip_file_path, :type => 'application/zip', :disposition => 'attachment', :stream => false
-    File.delete(@zip_file_path)
-  end  
+    begin
+      send_file @zip_file_path, :type => 'application/zip', :disposition => 'attachment', :stream => false
+      File.delete(@zip_file_path)
+    rescue  
+      if File.file?(@zip_file_path)
+        File.delete(@zip_file_path)
+      end  
+      puts "Error sending file"
+    end
+  end 
+  
+  def zip_filtered_single
+    @job = Job.find(params[:job_id])
+    unless User.current.admin? || @job.is_manager?
+      flash[:error] = "You are not authorized to view this section."
+  		redirect_to('/') and return
+  	end
+  	
+  	@job_applications = JobApplication.find(:all, :conditions => ["id in (?)", params[:job_applications]])
+  	
+    # create zip file of filtered results
+    @material_types = @job.application_material_types.split(',')
+    @file_name = @job.title.gsub(/[^a-zA-Z\d]/, '-')
+    @zip_file_path = "#{RAILS_ROOT}/tmp/#{@file_name}-filtered-materials.zip"
+    @ja_materials = Array.new
+    @ja_referrals = Array.new
+    filepaths = Array.new
+    counter = 1
+    material_id_hash = Hash.new
+
+    @applications = JobApplication.find(:all, :conditions => {:job_id => @job.id})
+
+    @job_applications.each do |app|
+      @ja_materials << JobApplicationMaterial.find(:first, :conditions => {:job_application_id => app.id})
+      unless @ja_referrals.nil?
+        @ja_referrals << JobApplicationReferral.find(:all, :conditions => {:job_application_id => app.id})
+      end  
+    end  
+
+    unless @ja_referrals.nil?
+      if @material_types.include?("Proposed Work")
+        @material_types.insert(@material_types.index("Proposed Work") + 1, "Referral")
+      elsif @material_types.include?("Cover Letter")
+        @material_types.insert(@material_types.index("Cover Letter") + 1, "Referral")
+      elsif @material_types.include?("Resume or CV")
+        @material_types.insert(@material_types.index("Resume or CV") + 1, "Referral")
+      else 
+        @material_types.insert(@material_types.index("Resume or CV") + 1, "Referral")
+      end  
+    end
+
+    @material_types.each do |material|
+      material_id_hash[material] = "%03d" % counter.to_s + "_" + material.gsub(/ /, '_')
+      counter = counter + 1
+    end
+
+    Zip::ZipFile.open(@zip_file_path, Zip::ZipFile::CREATE) do |zipfile|
+
+      unless @ja_materials.nil?
+        @ja_materials.each do |jam|
+          jam.attachments.each do |jama|
+            ext_name = File.extname("#{RAILS_ROOT}/files/" + jama.disk_filename)
+            new_file_name = "#{Applicant.find(JobApplication.find(jam.job_application_id).applicant_id).last_name}_#{Applicant.find(JobApplication.find(jam.job_application_id).applicant_id).first_name}_#{material_id_hash[jama.description]}_#{jam.job_application_id}#{ext_name}"
+            orig_file_path = "#{RAILS_ROOT}/files/" + jama.disk_filename
+            if File.exists?(orig_file_path)
+              orig_file_name = File.basename(orig_file_path)
+              if zipfile.find_entry(new_file_name)
+                zipfile.remove(new_file_name)
+              end
+              zipfile.get_output_stream(new_file_name) do |f|
+                input = File.open(orig_file_path)
+                data_to_copy = input.read()
+                f.write(data_to_copy)
+              end
+            else
+              puts "Warning: file #{orig_file_path} does not exist"
+            end
+          end    
+        end
+      end   
+
+      unless @ja_referrals.nil?
+        @ja_referrals.each do |jar|
+          jar.each do |ref|
+            ref.attachments.each do |jara|
+              ext_name = File.extname("#{RAILS_ROOT}/files/" + jara.disk_filename)
+              new_file_name = "#{Applicant.find(JobApplication.find(ref.job_application_id).applicant_id).last_name}_#{Applicant.find(JobApplication.find(ref.job_application_id).applicant_id).first_name}_#{material_id_hash[jama.description]}_#{ref.attachments.index(jara)+1}_#{ref.job_application_id}#{ext_name}"
+
+              orig_file_path = "#{RAILS_ROOT}/files/" + jara.disk_filename
+              if File.exists?(orig_file_path)
+                orig_file_name = File.basename(orig_file_path)
+                if zipfile.find_entry(new_file_name)
+                  zipfile.remove(new_file_name)
+                end
+                zipfile.get_output_stream(new_file_name) do |f|
+                  input = File.open(orig_file_path)
+                  data_to_copy = input.read()
+                  f.write(data_to_copy)
+                end
+              else
+                puts "Warning: file #{orig_file_path} does not exist"
+              end
+            end  
+          end    
+        end
+      end
+
+    end
+    begin
+      send_file @zip_file_path, :type => 'application/zip', :disposition => 'attachment', :stream => false
+      File.delete(@zip_file_path)
+    rescue  
+      if File.file?(@zip_file_path)
+        File.delete(@zip_file_path)
+      end  
+      puts "Error sending file"
+    end  
+  end 
+  
+  def filter_by_status
+    @job = Job.find(params[:job_id])
+    @apptracker = Apptracker.find(params[:apptracker_id])
+    unless User.current.admin? || @job.is_manager?
+      flash[:error] = "You are not authorized to view this section."
+  		redirect_to('/') and return
+  	end
+  	
+  	@job_application_custom_fields = @job.all_job_app_custom_fields
+    @applicant_fields = Applicant.column_names - ["id", "created_at", "updated_at"]
+    @custom = []
+    unless @job_application_custom_fields.empty?
+  		@job_application_custom_fields.each do |custom_field|
+  		  @custom << custom_field.name
+  		end
+  	end
+  	@columns = @applicant_fields + @custom
+  	@job_applications = []
+  	unless params[:review_status].blank?
+  	  @job_applications << JobApplication.find(:all, :conditions => {:job_id => params[:job_id], :review_status => params[:review_status]})
+  	end 
+  	unless params[:offer_status].blank?
+  	  @job_applications << JobApplication.find(:all, :conditions => {:job_id => params[:job_id], :offer_status => params[:offer_status]})
+  	end 
+  	@job_applications.flatten!
+  end
   
 end
